fix(gallery): get the right limit when the app starts
get the right limit when the app starts

MODULE-2-PART-4-mongodb
